Threw around compare functions, moved validity tests from loadFile to different functions (in order to facilitate for the absence of e.g. TLE)
