+++ examples/test/TestCase.h
@@ -3,6 +3,7 @@
 
 #include <string>
 #include <vector>
+#include <limits>
 
 class TestCase
 {
@@ -10,7 +11,7 @@ public:
 	TestCase();
 	~TestCase();
 
-	int loadFromFile(const char *filename);
+	void loadFromFile(const char *filename);
 
 	void getTLE(char *tle[2]);
 
@@ -20,6 +21,10 @@ public:
 
 	std::vector<std::vector<double> > &data() {return m_data;}
 
+	bool containsValidData();
+	bool containsValidQth();
+	bool containsValidTLE();
+
 private:
 	std::string m_tle[2];
 
@@ -28,7 +33,17 @@ private:
 	double m_qth_altitude;
 
 	std::vector<std::vector<double> > m_data;
-
 };
+
+bool fuzzyCompare(const double &x, const double &y, const double &epsilon = std::numeric_limits<double>::epsilon());
+
+/**
+ * Check whether input value lies within supplied boundary values. Order of boundary values is irrelevant.
+ * \param boundary_value_1 Boundary value 1
+ * \param boundary_value_2 Boundary value 2
+ * \param compared_value Compared value
+ * \return True if the compared value lies within the boundary values
+ **/
+bool fuzzyCompareWithBoundaries(const double &boundary_value_1, const double &boundary_value_2, const double &compared_value);
 
 #endif
